webapp/testing: specialize Tester
